refactor: Find Util
