+++ code/compiler/builder.node/common/index.mts
@@ -3,4 +3,7 @@ import { fs } from './fs.mjs';
 
 export { t, fs };
 export * from './libs.mjs';
-export * from './util.mjs';
+
+export { Util } from './util.mjs';
+export { FindUtil } from './util.Find.mjs';
+export { JsonUtil } from './util.Json.mjs';
