+++ script.ts/cmd.all.build.mts
@@ -13,7 +13,7 @@ const filter = (path: string) => {
   if (path.includes('code/samples/')) return false;
   return true;
 };
-let paths = await Builder.findProjectDirs({ filter, sort: 'DependencyGraph' });
+let paths = await Builder.Find.projectDirs({ filter, sort: 'DependencyGraph' });
 
 // Log module list.
 console.log();
